[Moved] Style class moved to individual file
