+++ HierarchyDecorator/Scripts/Editor/Style.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 249019516b184f040b3acc35ff0ee7d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
